Fixed null y values, added sorting
Created a second variable storing the dataset.

Use second dataset to replace this.data using a filter to ensure nothing on the y axis has a null value

Added a sort to ensure bars are ordered in descending order.

Changed the SVG height to be a constant, based on the initial dataset rather than the filtered one.